TST: Port some seqview and serialisation tests
